- Fix of "excess CPU bug" reported in fltk.bugs(not in STR) which was caused by not initializing no_items in Fl_Hash_Table - Fluid nis now synchronised with fltk-1.1x svn (r4300) (other sources not yet) - More fluid/style updates - More style updates (check button) - device dsp fix
git-svn-id: file:///fltk/svn/branches/branch-1.2@4311 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fluid/undo.cxx
@@ -31,8 +31,9 @@
 #include <FL/Fl_Preferences.H>
 #include "../src/flstring.h"
 #if defined(WIN32) && !defined(__CYGWIN__)
+#  include <io.h>
 #  include <windows.h>
-#  define getpid GetCurrentProcessId
+#  define getpid (int)GetCurrentProcessId
 #else
 #  include <unistd.h>
 #endif // WIN32 && !__CYGWIN__
